Revert "refactor(sl): check user ability before loading full video resource (#1140)"
This reverts commit e9529fbf06756c4df43be48aad1c0736d62bdaf8.
