+++ apps/epic-web/src/pages/tutorials/[module]/[section]/[lesson]/embed.tsx
@@ -21,16 +21,21 @@ import {ModuleProgressProvider} from '@skillrecordings/skill-lesson/video/module
 import {useRouter} from 'next/router'
 import pluralize from 'pluralize'
 import {trpc} from 'trpc/trpc.client'
-import {getCsrfToken, getProviders, useSession} from 'next-auth/react'
+import {getCsrfToken, getProviders, signIn, useSession} from 'next-auth/react'
 import {type LoginTemplateProps} from '@skillrecordings/ui/templates/login'
 import Spinner from 'components/spinner'
 import {Button} from '@skillrecordings/ui'
 import Link from 'next/link'
-import {createAppAbility} from '@skillrecordings/skill-lesson/utils/ability'
+import {
+  UserSchema,
+  createAppAbility,
+  defineRulesForPurchases,
+} from '@skillrecordings/skill-lesson/utils/ability'
+import {getToken} from 'next-auth/jwt'
 import {getSubscriberFromCookie} from '@skillrecordings/skill-lesson/utils/ck-subscriber-from-cookie'
+import {getProducts} from '@skillrecordings/skill-lesson/lib/products'
 import {getVideoResource} from '@skillrecordings/skill-lesson/lib/video-resources'
 import {Subscriber} from 'schemas/subscriber'
-import {defineAbilityRulesForResource} from '@skillrecordings/skill-lesson/utils/define-resource-ability-rules'
 
 export const getServerSideProps: GetServerSideProps = async (context) => {
   // resource
@@ -44,18 +49,43 @@ export const getServerSideProps: GetServerSideProps = async (context) => {
   const lesson = await getExercise(lessonSlug, false)
   const videoResourceId = lesson.videoResourceId
 
-  // rules
-  const abilityRules = await defineAbilityRulesForResource({
-    req: context.req as NextApiRequest,
-    moduleSlug,
-    sectionSlug,
-    lessonSlug,
-  })
-  const ability = createAppAbility(abilityRules || [])
-  const canShowVideo = ability.can('view', 'Content')
+  // ability rules
+  const country =
+    (context.req.headers['x-vercel-ip-country'] as string) ||
+    process.env.DEFAULT_COUNTRY ||
+    'US'
+
+  const token = await getToken({req: context.req})
   const convertkitSubscriber = await getSubscriberFromCookie(
     context.req as NextApiRequest,
   )
+
+  let purchasedModules = []
+
+  if (token) {
+    const user = UserSchema.parse(token)
+    const productsPurchased =
+      user.purchases?.map((purchase) => purchase.productId) || []
+    purchasedModules = await getProducts(productsPurchased)
+  }
+
+  const abilityRules = defineRulesForPurchases({
+    ...(token && {user: UserSchema.parse(token)}),
+    ...(convertkitSubscriber && {
+      subscriber: convertkitSubscriber,
+    }),
+    country,
+    module,
+    lesson,
+    section,
+    purchasedModules,
+    isSolution: lesson._type === 'solution',
+  })
+
+  const ability = createAppAbility(abilityRules || [])
+
+  // can show video
+  const canShowVideo = ability.can('view', 'Content')
 
   // full video resource
   const videoResource =
@@ -234,12 +264,7 @@ const Video: React.FC<
       status: videoResourceStatus,
       refetch: refetchVideoResource,
     } = trpc.videoResource.byId.useQuery(
-      {
-        id: Boolean(canShowVideo) ? videoResourceId : '',
-        moduleSlug: module.slug.current,
-        lessonSlug: lesson.slug,
-        sectionSlug: section?.slug,
-      },
+      {id: Boolean(canShowVideo) ? videoResourceId : ''},
       {
         refetchOnWindowFocus: true,
         staleTime: 5000,
