+++ packages/skill-lesson/hooks/use-video-resource.tsx
@@ -1,7 +1,6 @@
 import * as React from 'react'
 import {type VideoResource} from '../schemas/video-resource'
 import {trpcSkillLessons} from '../utils/trpc-skill-lessons'
-import {useLesson} from './use-lesson'
 
 type VideoResourceContextType = {
   videoResource?: VideoResource
@@ -22,15 +21,9 @@ export const VideoResourceProvider: React.FC<VideoResourceProviderProps> = ({
   videoResourceId,
   children,
 }) => {
-  const {module, section, lesson} = useLesson()
   const {data: videoResource, status} =
     trpcSkillLessons.videoResource.byId.useQuery(
-      {
-        id: videoResourceId,
-        moduleSlug: module?.slug.current,
-        lessonSlug: lesson.slug,
-        sectionSlug: section?.slug,
-      },
+      {id: videoResourceId},
       {
         refetchOnWindowFocus: false,
       },
@@ -38,7 +31,7 @@ export const VideoResourceProvider: React.FC<VideoResourceProviderProps> = ({
 
   const context = {
     videoResourceId,
-    videoResource: videoResource || undefined,
+    videoResource,
     loadingVideoResource: status === 'loading',
   }
   return (
